+++ src/gc/gcexpr.c
@@ -96,7 +96,7 @@ PExpr *NewLogical(Pgc *gc, PExpr *left, Token *op, PExpr *right) {
 }
 
 PExpr *NewCallExpr(
-    Pgc *gc, Token *op, PExpr *callee, PExpr **args, size_t count
+    Pgc *gc, Token *op, PExpr *callee, PExpr **args, u64 count
 ) {
     PExpr *e = NewExpr(gc, EXPR_CALL, op);
     if (e == NULL) {
@@ -109,7 +109,7 @@ PExpr *NewCallExpr(
     return e;
 }
 
-PExpr *NewArrayExpr(Pgc *gc, Token *op, PExpr **items, size_t count) {
+PExpr *NewArrayExpr(Pgc *gc, Token *op, PExpr **items, u64 count) {
     PExpr *e = NewExpr(gc, EXPR_ARRAY, op);
     if (e == NULL) {
         return NULL;
@@ -120,7 +120,7 @@ PExpr *NewArrayExpr(Pgc *gc, Token *op, PExpr **items, size_t count) {
     return e;
 }
 
-PExpr *NewMapExpr(Pgc *gc, Token *op, PExpr **items, size_t count) {
+PExpr *NewMapExpr(Pgc *gc, Token *op, PExpr **items, u64 count) {
     PExpr *e = NewExpr(gc, EXPR_MAP, op);
     if (e == NULL) {
         return NULL;
@@ -176,8 +176,8 @@ void FreeExpr(Pgc *gc, PExpr *e) {
     switch (e->type) {
         case EXPR_CALL: {
             FreeExpr(gc, e->exp.ECall.callee);
-            size_t count = e->exp.ECall.argCount;
-            for (size_t i = 0; i < count; i++) {
+            u64 count = e->exp.ECall.argCount;
+            for (u64 i = 0; i < count; i++) {
                 PExpr *ex = arrpop(e->exp.ECall.args);
                 FreeExpr(gc, ex);
             }
@@ -233,8 +233,8 @@ void FreeExpr(Pgc *gc, PExpr *e) {
             break;
         }
         case EXPR_ARRAY: {
-            size_t count = e->exp.EArray.count;
-            for (size_t i = 0; i < count; i++) {
+            u64 count = e->exp.EArray.count;
+            for (u64 i = 0; i < count; i++) {
                 FreeExpr(gc, arrpop(e->exp.EArray.items));
             }
             arrfree(e->exp.EArray.items);
@@ -244,7 +244,7 @@ void FreeExpr(Pgc *gc, PExpr *e) {
         }
         case EXPR_MAP: {
             struct EMap *map = &e->exp.EMap;
-            for (size_t i = 0; i < map->count; i++) {
+            for (u64 i = 0; i < map->count; i++) {
                 FreeExpr(gc, arrpop(map->etable));
             }
